pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "workflow-app"
version = "0.1.0"
description = "Workflow App — Claude Code pipeline orchestrator"
requires-python = ">=3.10"
dependencies = [
    "PySide6>=6.6.0",
    "SQLAlchemy>=2.0.0",
    "alembic>=1.13.0",
    "python-statemachine>=2.0.0",
    "pyte>=0.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-qt>=4.0",
    "pytest-timeout>=2.0",
    "pytest-cov>=5.0",
    "ruff>=0.4.0",
    "mypy>=1.8.0",
    "mutmut>=2.4",
]

[project.scripts]
workflow-app = "workflow_app.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/workflow_app"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.hatch.envs.default.scripts]
db-seed = "python3 scripts/seed.py"
db-reset = "python3 scripts/seed.py --reset"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]
ignore = ["E501"]  # line too long handled by formatter

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
exclude = ["tests/", "scripts/"]

[tool.mutmut]
paths_to_mutate = "src/workflow_app/"
backup = false
runner = "python -m pytest tests/ -x --timeout=30 --ignore=tests/performance -q"
tests_dir = "tests/"
# Excluir módulos UI (PySide6) — alto custo de mutação, baixo valor de lógica pura
# Focar em: domain, pipeline, sdk, core, db, history, errors
dict_synonyms = "Mutation, mutation"
